[⚙️ chore] ProfileShortcut 컴포넌트 분리 #2
