+++ src/pages/profile/index.tsx
@@ -1,9 +1,9 @@
 import styled from '@emotion/styled';
 import Image from 'next/image';
 import Link from 'next/link';
-import { useRouter } from 'next/router';
 import Layout from '@/components/common/Layout';
 import JOBTITLE from '@/constants/jobTitle';
+import ProfileShortcut from '@/components/profile/ProfileShortcut';
 
 export default function Profile() {
   const profileInfo = {
@@ -12,15 +12,6 @@ export default function Profile() {
     githubId: 'hello_world',
   };
   const isLogined = true;
-
-  const router = useRouter();
-
-  const goToGithub = () => {
-    router.push('/profile/github').catch((error) => {
-      console.error('페이지 이동에 실패하였습니다.', error);
-    });
-  };
-
   return (
     <Layout
       noHeader
@@ -37,49 +28,7 @@ export default function Profile() {
       </SNotificationBtn>
       <SProfileWrapper>
         <h2 className="a11yHidden">프로필</h2>
-        <SProfileShortcutWrapper>
-          <h3 className="a11yHidden">프로필 인사말</h3>
-          <SProfileGreetingWrapper>
-            {isLogined ? (
-              <p>
-                <span>{profileInfo.nickName}</span>&nbsp;{profileInfo.jobTitle}
-              </p>
-            ) : (
-              <Link href="/">
-                <p>
-                  <span>로그인</span>&nbsp;혹은&nbsp;<span>회원가입</span>
-                  <Image
-                    src="/icons/icon-down-arrow.svg"
-                    alt="화살표 아이콘"
-                    width={34}
-                    height={34}
-                    style={{ transform: 'rotate(270deg)' }}
-                  />
-                </p>
-              </Link>
-            )}
-            <p>개발자님 오늘도 화이팅하세요!</p>
-          </SProfileGreetingWrapper>
-          {isLogined && (
-            <SGithubIdBtn
-              type="button"
-              onClick={goToGithub}
-              isGithub={!!profileInfo.githubId}
-            >
-              <Image
-                src="/icons/icon-github-logo.svg"
-                alt="깃허브 로고"
-                width={18}
-                height={18}
-              />
-              <p>
-                {isLogined && profileInfo.githubId
-                  ? profileInfo.githubId
-                  : '깃허브 연동하기'}
-              </p>
-            </SGithubIdBtn>
-          )}
-        </SProfileShortcutWrapper>
+        <ProfileShortcut isLogined={isLogined} profileInfo={profileInfo} />
         <div>
           <h3>챌린지 기록</h3>
           <ul>
@@ -230,7 +179,6 @@ export default function Profile() {
             <p>1.0.0</p>
           </div>
         </SProfileEtc>
-
         <SwithdrawalBtnWrapper>
           {isLogined && <button type="button">회원 탈퇴</button>}
         </SwithdrawalBtnWrapper>
@@ -286,55 +234,6 @@ const SProfileActiveMenuWrapper = styled(SPropfileBaseMenuWrapper)`
       isLogined ? theme.color.gray_3c : theme.color.gray_bf};
     cursor: ${({ isLogined }) => (isLogined ? 'pointer' : 'not-allowed')};
   }
-`;
-
-const SProfileShortcutWrapper = styled.div`
-  margin-top: 3.5rem;
-  display: flex;
-  flex-flow: column nowrap;
-  justify-content: center;
-  gap: 1.5rem;
-
-  &::after {
-    content: ' ';
-    width: calc(100% + 2.5rem);
-    border-bottom: 6px solid ${({ theme }) => theme.color.gray_ec};
-    position: relative;
-    left: -1.25rem;
-  }
-`;
-
-const SProfileGreetingWrapper = styled.div`
-  width: 335px;
-  height: 70px;
-
-  & p {
-    font-size: ${({ theme }) => theme.fontSize.headline1};
-    font-weight: ${({ theme }) => theme.fontWeight.headline1};
-    display: inline-flex;
-    flex-flow: row nowrap;
-    align-items: center;
-  }
-
-  & span {
-    color: ${({ theme }) => theme.color.normal};
-  }
-`;
-
-const SGithubIdBtn = styled.button<{ isGithub: boolean }>`
-  width: ${({ isGithub }) => (isGithub ? '118px' : '135px')};
-  height: 28px;
-  border-radius: 16px;
-  background-color: ${({ theme }) => theme.color.gray_83};
-  display: flex;
-  flex-flow: row nowrap;
-  align-items: center;
-  justify-content: space-between;
-  gap: 8px;
-  padding: 0 10px;
-  color: ${({ theme }) => theme.color.gray_f9};
-  font-size: ${({ theme }) => theme.fontSize.body4};
-  font-weight: ${({ theme }) => theme.fontWeight.body4};
 `;
 
 const SProfileEtc = styled.div`
